% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_xl.R
\name{left_xl}
\alias{left_xl}
\title{left_xl function
This function replicates the LEFT function in Excel and is utilised for left trimming
of character strings}
\usage{
left_xl(text, num_char = 0)
}
\arguments{
\item{text}{The text you want to LEFT trim}

\item{num_char}{The number of characters your want to trim by}
}
\value{

}
\description{
left_xl function
This function replicates the LEFT function in Excel and is utilised for left trimming
of character strings
}
\examples{
left_xl(text= "This is some example text", num_char = 4)

}
