% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len_xl.R
\name{len_xl}
\alias{len_xl}
\title{len_xl function}
\usage{
len_xl(text, ...)
}
\arguments{
\item{text}{The text to get the numeric length of}

\item{...}{Function forwarding to work with the base nchar method}
}
\description{
This function replicates the LEN function in Excel and is utilised for finding the length of character strings.
}
\examples{
len_xl("Guess the length of me!")


}
