% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_xl.R
\name{mid_xl}
\alias{mid_xl}
\title{mid_xl function}
\usage{
mid_xl(text, start_num = 1, num_char = 0)
}
\arguments{
\item{text}{The text you want to MID trim}

\item{start_num}{The start number to start the trim. This needs to be numeric.}

\item{num_char}{The number of characters your want to trim by. This field needs to be numeric.}
}
\value{
The extracted text between the start_num and the num_char to produce a sub string result.
}
\description{
This function replicates the MID function in Excel and is utilised for left trimming
of character strings.
}
\details{
This has been included as a convenience function for working with text and string data.
}
\examples{
mid_xl(text= "This is some example text", start_num = 6, num_char = 10)

\donttest{
mid_xl(text= "This is some example text", start_num = 6, num_char = 10)
}

}
