% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSafelyCodeList.R
\name{openSafelyCodeList}
\alias{openSafelyCodeList}
\title{openSafelyCodeList function}
\usage{
openSafelyCodeList(list_name, version = "", ...)
}
\arguments{
\item{list_name}{The code list ID from \url{https://codelists.opensafely.org/} for which to return the National table of elements, for example
"opensafely/ace-inhibitor-medications"}

\item{version}{The version of the code list if not the most recent}

\item{...}{Function forwarding to parent function to pass additional arguments to function (e.g. title, add_zero_prefix)}
}
\value{
A tibble (class data.frame) output from the results of the web scrape
\itemize{
  \item result - the extracted code list from the full list tab of the Code Lists page
  \item DictType - defaults to Not Specified if nothing passed, however allows for custom dictionary / data frame tags to be created
  \item DttmExtracted - a date and time stamp
}
}
\description{
This function uses the \code{\link{tableR}} parent function to return a table of elements, specifically from the OpenSafely Code List
\url{https://codelists.opensafely.org/}
}
\examples{
openSafelyCodeList("opensafely/ace-inhibitor-medications")
# Returns a tibble with the Code and Description from https://codelists.opensafely.org/codelist/opensafely/ace-inhibitor-medications using the most recent version
# with Dict_Type added and a Date and Time the dictionary was extracted.

openSafelyCodeList("opensafely/ace-inhibitor-medications", "2020-05-19")
# Returns a tibble with the Code and Description from https://codelists.opensafely.org/codelist/opensafely/ace-inhibitor-medications/2020-05-19 version
# with Dict_Type added and a Date and Time the dictionary was extracted.

}
