% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_xl.R
\name{right_xl}
\alias{right_xl}
\title{right_xl function}
\usage{
right_xl(text, num_char = 0)
}
\arguments{
\item{text}{The text you want to RIGHT trim}

\item{num_char}{The number of characters your want to trim by. This field needs to be numeric.}
}
\value{

}
\description{
This function replicates the RIGHT function in Excel and is utilised for right trimming
of character strings.
}
\details{
This has been included as a convenience function for working with text and string data.
}
\examples{
right_xl(text= "This is some example text", num_char = 10)

\donttest{
right_xl(text= "This is some example text", num_char = 10)
}

}
