% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeR.R
\name{scrapeR}
\alias{scrapeR}
\title{ScrapeR - scrape web information with scrapeR}
\usage{
scrapeR(url, xpath, ...)
}
\arguments{
\item{url}{Website address to connect to}

\item{xpath}{Xpath obtained through inspecting the individual HTML elements}

\item{...}{Function to pass additional function forwarding options}
}
\value{
Returns the results of the scraping operation and the relevant fields from the html table - the xpath should make reference to an html table, otherwise an error is returned advising the user to check the xpath and url are correct.
}
\description{
Takes the url and xpath and scrapes HTML table elements from a website.
}
\details{
This function is specifically designed to work with HTML tables and x path links through to direct HTML elements.
The function is versatile and can be used on any URL where an xpath can be obtained through the URL and HTML inspection process.
}
