% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableR.R
\name{tableR}
\alias{tableR}
\title{tableR function}
\usage{
tableR(url, xpath, title, add_zero_prefix = FALSE, ...)
}
\arguments{
\item{url}{The URL of the website to scrape the table element from}

\item{xpath}{The unique xpath of the HTML element to be scraped}

\item{title}{Unique name for the relevant HTML table that has been scraped}

\item{add_zero_prefix}{Adds zero prefixes to certain codes that get converted by native functions}

\item{...}{Function forwarding to parent function to pass additional arguments to function}
}
\value{

}
\description{
This function uses the scapeR parent function to return a table of elements
}
\examples{
permitted_codes <- tableR(url= "https://datadictionary.nhs.uk/data_elements/abdominal_x-ray_performed_reason.html", xpath = '//*[@id="element_abdominal_x-ray_performed_reason.permitted_codes"]/div/div/table', title = "x-ray performed", add_zero_prefix = TRUE)
}
