% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_data_elements.R
\name{nhs_data_elements}
\alias{nhs_data_elements}
\title{NHS data elements method}
\usage{
nhs_data_elements()
}
\value{
A tibble (class data frame) with the results of scraping the NHS Data Dictionary website for the data elements look ups, if no return this will produce an appropriate informational message.
\itemize{
  \item link_name - the name of the scraped link. This relates to the actual name of the data element from the NHS Data Dictionary.
  \item url - the url passed to the parameter
  \item full_url - the full url of where the data element is on the NHS Data Dictionary website
  \item xpath_nat_code - utilises the element in the website and appends the link_short - to pull back only national codes from the dictionary site. NOTE: not all of the returns will have national code tables.
  \item xpath_also_known - pulls back the data dictionary elements alias table - this will be available for all data elements

}
}
\description{
Searches all the data elements in the data element index of the NHS data dictionary and returns the links.
}
\details{
This function has no input parameters and returns the
}
\examples{
nhs_data_lookup <- nhs_data_elements()
head(nhs_data_lookup, 10)
}
