% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_data_elements.R
\name{nhs_data_elements}
\alias{nhs_data_elements}
\title{NHS data elements method}
\usage{
nhs_data_elements()
}
\description{
Searches all the data elements in the data element index of the NHS data dictionary and returns the links.
}
\details{
This function has no input parameters and returns the
}
\examples{
nhs_data_lookup <- nhs_data_elements()
head(nhs_data_lookup, 10)
}
