% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpathTextR.R
\name{xpathTextR}
\alias{xpathTextR}
\title{xpathTextR function}
\usage{
xpathTextR(url, xpath)
}
\arguments{
\item{url}{The link for the website}

\item{xpath}{The xpath string derived by using the Inspect functionality in a web browser.}
}
\value{

}
\description{
Returns xpath text from websites and can be used to access specific HTML nodes
}
