% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpathTextR.R
\name{xpathTextR}
\alias{xpathTextR}
\title{xpathTextR function}
\usage{
xpathTextR(url, xpath, ssl_needed = FALSE)
}
\arguments{
\item{url}{The link for the website}

\item{xpath}{The xpath string derived by using the Inspect functionality in a web browser.}

\item{ssl_needed}{Default - FALSE - Boolean to indicate whether to need a SSL certificate}
}
\value{
A list with the results of scraping the specific xpath element
\itemize{
  \item result - the extracted text from the website element that has been scraped
  \item website_passed - a copy of the input url for the website
  \item html_node_result - returns the extracted html node result
  \item datetime_access - returns a timestamp of when the results of the scraping operation have been completed
  \item person_accessed - retrieves the system environment stored username and domain - this is concatenated together to form a mixed charatcer string

}
}
\description{
Returns xpath text from websites and can be used to access specific HTML nodes
}
